+++ core-algo/temp/graph.cpp
@@ -1,121 +1,10 @@
 #include<iostream>
-
 #include<unordered_map>
 #include<vector>
 #include<cstdio>
-
 #include<assert.h>
 
 #include "graph.hpp"
-
-//overriding stds hash function for the VertexData class
-
-namespace std {
-  template <>
-  struct hash<VertexData>
-  {
-    std::size_t operator()(const VertexData& v_data) const
-    {
-        using std::size_t;
-        using std::hash;
-        // Compute individual hash values for first,
-        // second and combine them using XOR and bit shifting:
-        pair<double, double> geolocation = v_data.get_geolocation();
-        return (hash<double>()(geolocation.first) ^ (hash<double>()(geolocation.second) << 1));
-    }
-  };
-
-}
-
-// Edge Class implementation
-
-Edge::Edge(std::pair<Vertex*, Vertex*> vertices, EdgeWeight weight){
-    this->vertices = vertices;
-    this->weight = weight;
-}
-
-Edge::Edge(){
-    this->vertices.first = nullptr;
-    this->vertices.second = nullptr;
-    this->weight = -1; // placeholder value for weight
-}
-
-void Edge::set_weight(EdgeWeight weight){
-    this->weight = weight;
-}
-
-void Edge::set_vertices(std::pair<Vertex*, Vertex*> vertices){
-    this->vertices = vertices;
-}
-/*
-class VertexData{
-    private:
-        std::pair<double, double> geolocation;
-
-        // the geolocation uniquely defines a vertex, i.e. two vertices with the same geolocations has to be equal
-        // TODO: think about the information we might want to hold in a vertex such as 
-        // order ids, address information, etc.
-
-    public:
-        VertexData(std::pair<double, double> geolocation);
-        VertexData();
-        std::pair<double, double> get_geolocation() const;
-        bool operator==(VertexData v_data);
-
-
-};
-*/
-std::pair<Vertex*, Vertex*> Edge::get_vertices(){
-    return this->vertices;
-}
-
-EdgeWeight Edge::get_weight(){
-    return this->weight;
-}
-
-// VertexData Class Implementation
-
-VertexData::VertexData(){
-    this->geolocation = std::make_pair(-1,-1); //placeholder values 
-}
-
-VertexData::VertexData(std::pair<double, double> geolocation){
-    this->geolocation = geolocation;
-}
-
-std::pair<double, double> VertexData::get_geolocation() const{
-    return this->geolocation;
-}
-
-bool VertexData::operator==(VertexData& v_data){
-    return v_data.get_geolocation() == this->geolocation;
-}
-
-
-// Vertex Class Implementaton
-
-Vertex::Vertex(){
-    this->v_data = VertexData();
-    this->v_index = -1; //placeholder value for index
-}
-
-Vertex::Vertex(VertexData v_data, int v_index){
-    this->v_data = v_data;
-    this->v_index = v_index;
-}
-
-int Vertex::get_index(){
-    return this->v_index;
-}
-
-VertexData Vertex::get_data(){
-    return this->v_data;
-}
-
-bool Vertex::operator==(Vertex& v2){
-    return v2.get_data() == this->v_data;
-}
-
 
 // Graph Class Implementation
